remote_server/requirements.txt
flask>=2.0.0,<3.0.0
flask-cors>=4.0.0,<5.0.0
flask-socketio>=5.3.0,<6.0.0
python-engineio>=4.0.0
python-socketio>=5.0.0
gevent
gevent-websocket
gunicorn
eventlet
python-dotenv
requests>=2.25.0